postEditのクラス整理
